backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
requests>=2.31.0
python-multipart>=0.0.9
pydantic>=2.0.0
python-dotenv>=1.0.1
